+++ entrypoint.sh
@@ -1,11 +1,7 @@
 #!/bin/sh
 
-set -eux
+set -eu
 
-# alternative: https://pkgs.alpinelinux.org/contents?file=envsubst&path=&name=&branch=v3.12
-sed -e "s#{version}#$VERSION#" \
-    -e "s#{virtual_port}#$VIRTUAL_PORT#" \
-    -e "s#{target}#$TARGET#" \
-    /torrc.template >/tmp/torrc
+envsubst </torrc.template >/tmp/torrc
 
 exec "$@"